Created a 3d space & a box to interact with.
Was based off of the tutorials at: wiki.jmonkeyengine.org

The box can be pushed, or shot and the game will mark where you shot in
space with a red sphere. The physics is a bit wonkey and just pulled from
a tutorial, but is editable at least somewhat. This is probably my last
commit until our meeting later today. It certainly shows proof of concept
for interacting with 3d objects.