added door proof of concept.
Now, there is a box (aka, button) and a vertical rectangle (aka, door).
When the box is shot, it changes color, and also removes the door from the
game space. This is at least a proof of concept for the user interacting
with objects indirectly.

We will need to work on map generation to allow our door to clip through
other objects, and not be affected by physics, but still have our player
collide with it.