Added code to display button, and setup for collision.
A majority of the code is copied and pasted from userInteraction, but it
works.

The code will need to be refactored to fit inside the classes we have
made. Along with the level generation needs to be implemented as well.
Having a space to place the objects is helpful to realise what we need to
be in what class.