Button.java is working as intended, collision still needs work.
Client.java now has a ghost object attached to the player to show
collision between the player and the button. Even though some collisions
are being detected, it is not wokring as intended.

Button.java is working as intended, and will be considered 'done' for the
time being. Player.java, level generation, and proper collision detection
are the top priorities.