cleaned up player & button classes, and worked on client.java a bit.
In order to have trigger methods for our classes, we may want to have
entity extend spatial. Right now, we can only get the geomentries that are
triggered. we may be able to have a check in every instance
triggerGeometry.equals(this.geometry) or something like that, but it'd
probably be neater to just extend spatial/geometry?