Players now are set to audio/visual player and render differently.
The first player to connect is always the audio player, the 2nd is the
visual.

The way the rendering works is:
	If I am audio player  - render my model for the visual player.
	If I am visual player - render my particles for the audio player.

This could be switched to have the client handle the rendering of
everything instead of the player itself, which might be needed in the
future anyways. since buttons need the same instances on both clients, but will
have different renders

I also need to have the particles only render on movement/trigger.