Started adding code for the InteractableEntity's messages.
The entitiy will have a list of things it can trigger, probably given at
it's creation. (it would be nice to be able to set after the fact, but
might be more complicated). When the entity is triggered, it will send a
message contaning the state of the object (T/F), along with a list of
indexes of objects that will be set. Then the object updates itself,
followed by all indexes it receives in the message.

Once this is working, making the client render objects differently should
be fairly simple to do. But we need to be able to register user's
interaction on both sides first.