Wrote a tiny bit of code for Boxes.
The way I see this working, Player has a reference to a box. if he shoots a box, he can 'pick it up' (store the reference & remove from rootNode). He can also shoot a button, if he has a box, he can drop it on the button.

The only things we should need to implement are the actual box (collision & geometry) and saving it to the level. We wont handle any sort of pushing or rotating, and the player can only place a box on a button. EZPZ